Request fd sets directly instead of by their events
When --use-select is specified, --in, --out and --pri
request readfds, writefds and exeptfds, respectively.
Otherwise, they request POLLIN, POLLOUT and POLLPRI.

In select(), events cannot be requested individually, they
are requested in groups (the fd sets) as follows:

- readfds:  POLLPRI | POLLERR | POLLIN  | POLLHUP
- writefds: POLLPRI | POLLERR | POLLOUT
- exeptfds: POLLPRI

Before this commit, an fd set was requested only if it
internally requested the POLL* events asked by user: POLLPRI
(--pollpri), POLLERR (--pollerr), POLLIN (--pollin) and
POLLOUT (--pollout).

If this change is not done, requesting POLLPRI results in
requesting all fd sets, then it is not possible to request
only exeptfds.

Now the fd sets are requested directly by the options.

The help text is clarified.

This commit reverts:

- commit c21bdfe422533e63f5ce ("Request readfds and writefds
  on POLLPRI").
- commit 8be9f38f39128f906fe4 ("Request error condition on
  select()").

Signed-off-by: Ricardo Biehl Pasquali <pasqualirb@gmail.com>